+++ pybind/extern.hpp
@@ -755,6 +755,13 @@ void instantiate_manager(py::module &m, const std::string &dtypestr) {
                &manager_type::to_string, py::const_))
       .def("insert_and_map", &manager_type::insert_and_map)
       .def("insert_field", &manager_type::insert_field)
+      .def("field_to_sparse_insert_and_map",
+           &manager_type::field_to_sparse_insert_and_map)
+      .def("exists_field_to_sparse",
+           py::overload_cast<minkowski::CoordinateMapKey const *,
+                             minkowski::CoordinateMapKey const *>(
+               &manager_type::exists_field_to_sparse, py::const_))
+      .def("get_field_to_sparse_map", &manager_type::get_field_to_sparse_map)
       .def("stride", &manager_type::py_stride)
       .def("origin", &manager_type::py_origin)
       .def("get_coordinates", &manager_type::get_coordinates)